chore(storage): move gRPC Writer code to own file
Part I of splitting out chunks of grpc_client.go
