+++ storage/grpc_client.go
@@ -22,7 +22,6 @@ import (
 	"hash/crc32"
 	"io"
 	"log"
-	"net/url"
 	"os"
 	"sync"
 
@@ -31,7 +30,6 @@ import (
 	gapic "cloud.google.com/go/storage/internal/apiv2"
 	"cloud.google.com/go/storage/internal/apiv2/storagepb"
 	"github.com/googleapis/gax-go/v2"
-	"google.golang.org/api/googleapi"
 	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
 	"google.golang.org/api/option/internaloption"
@@ -56,12 +54,6 @@ const (
 	//
 	// This is only used for the gRPC client.
 	defaultConnPoolSize = 1
-
-	// maxPerMessageWriteSize is the maximum amount of content that can be sent
-	// per WriteObjectRequest message. A buffer reaching this amount will
-	// precipitate a flush of the buffer. It is only used by the gRPC Writer
-	// implementation.
-	maxPerMessageWriteSize int = int(storagepb.ServiceConstants_MAX_WRITE_CHUNK_BYTES)
 
 	// globalProjectAlias is the project ID alias used for global buckets.
 	//
@@ -1736,116 +1728,6 @@ func (c *grpcStorageClient) NewRangeReader(ctx context.Context, params *newRange
 	return r, nil
 }
 
-func (c *grpcStorageClient) OpenWriter(params *openWriterParams, opts ...storageOption) (*io.PipeWriter, error) {
-	var offset int64
-	errorf := params.setError
-	setObj := params.setObj
-	setFlush := params.setFlush
-	pr, pw := io.Pipe()
-
-	s := callSettings(c.settings, opts...)
-
-	retryDeadline := defaultWriteChunkRetryDeadline
-	if params.chunkRetryDeadline != 0 {
-		retryDeadline = params.chunkRetryDeadline
-	}
-	if s.retry == nil {
-		s.retry = defaultRetry.clone()
-	}
-	s.retry.maxRetryDuration = retryDeadline
-
-	// Set Flush func for use by exported Writer.Flush.
-	var gw *gRPCWriter
-	setFlush(func() (int64, error) {
-		return gw.flush()
-	})
-	gw, err := newGRPCWriter(c, s, params, pr, pw, params.setPipeWriter)
-	if err != nil {
-		errorf(err)
-		pr.CloseWithError(err)
-		close(params.donec)
-		return nil, err
-	}
-
-	var o *storagepb.Object
-
-	// If we are taking over an appendable object, send the first message here
-	// to get the append offset.
-	if params.appendGen > 0 {
-		// Create the buffer sender. This opens a stream and blocks until we
-		// get a response that tells us what offset to write from.
-		wbs, err := gw.newGRPCAppendTakeoverWriteBufferSender(params.ctx)
-		if err != nil {
-			return nil, fmt.Errorf("storage: creating buffer sender: %w", err)
-		}
-		// Propagate append offset to caller and buffer sending logic below.
-		params.setTakeoverOffset(wbs.takeoverOffset)
-		offset = wbs.takeoverOffset
-		gw.streamSender = wbs
-		o = wbs.objResource
-		setObj(newObjectFromProto(o))
-	}
-
-	// This function reads the data sent to the pipe and sends sets of messages
-	// on the gRPC client-stream as the buffer is filled.
-	go func() {
-		err := func() error {
-			// Unless the user told us the content type, we have to determine it from
-			// the first read.
-			if params.attrs.ContentType == "" && !params.forceEmptyContentType {
-				gw.reader, gw.spec.Resource.ContentType = gax.DetermineContentType(gw.reader)
-			}
-
-			// Loop until there is an error or the Object has been finalized.
-			for {
-				// Note: This blocks until either the buffer is full or EOF is read.
-				recvd, doneReading, err := gw.read()
-				if err != nil {
-					return err
-				}
-
-				uploadBuff := func(ctx context.Context) error {
-					obj, err := gw.uploadBuffer(ctx, recvd, offset, doneReading)
-					if obj != nil {
-						o = obj
-						setObj(newObjectFromProto(o))
-					}
-					return err
-				}
-
-				// Add routing headers to the context metadata for single-shot and resumable
-				// writes. Append writes need to set this at a lower level to pass the routing
-				// token.
-				bctx := gw.ctx
-				if !gw.append {
-					bctx = bucketContext(bctx, gw.bucket)
-				}
-				err = run(bctx, uploadBuff, gw.settings.retry, s.idempotent)
-				if err != nil {
-					return err
-				}
-				offset += int64(recvd)
-
-				// When we are done reading data without errors, set the object and
-				// finish.
-				if doneReading {
-					// Build Object from server's response.
-					setObj(newObjectFromProto(o))
-					return nil
-				}
-			}
-		}()
-
-		// These calls are still valid if err is nil
-		err = checkCanceled(err)
-		errorf(err)
-		pr.CloseWithError(err)
-		close(params.donec)
-	}()
-
-	return pw, nil
-}
-
 // IAM methods.
 
 func (c *grpcStorageClient) GetIamPolicy(ctx context.Context, resource string, version int32, opts ...storageOption) (*iampb.Policy, error) {
@@ -2612,468 +2494,4 @@ func (r *gRPCReader) reopenStream() error {
 	r.currMsg = res.decoder
 	r.cancel = cancel
 	return nil
-}
-
-func newGRPCWriter(c *grpcStorageClient, s *settings, params *openWriterParams, r io.Reader, pw *io.PipeWriter, setPipeWriter func(*io.PipeWriter)) (*gRPCWriter, error) {
-	if params.attrs.Retention != nil {
-		// TO-DO: remove once ObjectRetention is available - see b/308194853
-		return nil, status.Errorf(codes.Unimplemented, "storage: object retention is not supported in gRPC")
-	}
-
-	size := googleapi.MinUploadChunkSize
-	// A completely bufferless upload (params.chunkSize <= 0) is not possible in
-	// gRPC because the buffer must be provided to the message. Use the minimum
-	// size possible.
-	if params.chunkSize > 0 {
-		size = params.chunkSize
-	}
-
-	// Round up chunksize to nearest 256KiB
-	if size%googleapi.MinUploadChunkSize != 0 {
-		size += googleapi.MinUploadChunkSize - (size % googleapi.MinUploadChunkSize)
-	}
-
-	if s.userProject != "" {
-		params.ctx = setUserProjectMetadata(params.ctx, s.userProject)
-	}
-
-	spec := &storagepb.WriteObjectSpec{
-		Resource:   params.attrs.toProtoObject(params.bucket),
-		Appendable: proto.Bool(params.append),
-	}
-	var appendSpec *storagepb.AppendObjectSpec
-	if params.appendGen > 0 {
-		appendSpec = &storagepb.AppendObjectSpec{
-			Bucket:     bucketResourceName(globalProjectAlias, params.bucket),
-			Object:     params.attrs.Name,
-			Generation: params.appendGen,
-		}
-	}
-	// WriteObject doesn't support the generation condition, so use default.
-	if err := applyCondsProto("WriteObject", defaultGen, params.conds, spec); err != nil {
-		return nil, err
-	}
-
-	return &gRPCWriter{
-		buf:                   make([]byte, size),
-		c:                     c,
-		ctx:                   params.ctx,
-		reader:                r,
-		pw:                    pw,
-		bucket:                params.bucket,
-		attrs:                 params.attrs,
-		conds:                 params.conds,
-		spec:                  spec,
-		appendSpec:            appendSpec,
-		encryptionKey:         params.encryptionKey,
-		settings:              s,
-		progress:              params.progress,
-		setSize:               params.setSize,
-		sendCRC32C:            params.sendCRC32C,
-		forceOneShot:          params.chunkSize <= 0,
-		forceEmptyContentType: params.forceEmptyContentType,
-		append:                params.append,
-		finalizeOnClose:       params.finalizeOnClose,
-		setPipeWriter:         setPipeWriter,
-		flushComplete:         make(chan int64),
-	}, nil
-}
-
-// gRPCWriter is a wrapper around the the gRPC client-stream API that manages
-// sending chunks of data provided by the user over the stream.
-type gRPCWriter struct {
-	c             *grpcStorageClient
-	buf           []byte
-	reader        io.Reader
-	pw            *io.PipeWriter
-	setPipeWriter func(*io.PipeWriter) // used to set in parent storage.Writer
-
-	ctx context.Context
-
-	bucket        string
-	attrs         *ObjectAttrs
-	conds         *Conditions
-	spec          *storagepb.WriteObjectSpec
-	appendSpec    *storagepb.AppendObjectSpec
-	encryptionKey []byte
-	settings      *settings
-	progress      func(int64)
-	setSize       func(int64)
-
-	sendCRC32C            bool
-	forceOneShot          bool
-	forceEmptyContentType bool
-	append                bool
-	finalizeOnClose       bool
-
-	streamSender    gRPCBidiWriteBufferSender
-	flushInProgress bool       // true when the pipe is being recreated for a flush.
-	flushComplete   chan int64 // use to signal back to flush call that flush to server was completed.
-}
-
-func bucketContext(ctx context.Context, bucket string) context.Context {
-	hds := []string{"x-goog-request-params", fmt.Sprintf("bucket=projects/_/buckets/%s", url.QueryEscape(bucket))}
-	return gax.InsertMetadataIntoOutgoingContext(ctx, hds...)
-}
-
-// drainInboundStream calls stream.Recv() repeatedly until an error is returned.
-// It returns the last Resource received on the stream, or nil if no Resource
-// was returned. drainInboundStream always returns a non-nil error. io.EOF
-// indicates all messages were successfully read.
-func drainInboundStream(stream storagepb.Storage_BidiWriteObjectClient) (object *storagepb.Object, err error) {
-	for err == nil {
-		var resp *storagepb.BidiWriteObjectResponse
-		resp, err = stream.Recv()
-		// GetResource() returns nil on a nil response
-		if resp.GetResource() != nil {
-			object = resp.GetResource()
-		}
-	}
-	return object, err
-}
-
-func bidiWriteObjectRequest(buf []byte, offset int64, flush, finishWrite bool) *storagepb.BidiWriteObjectRequest {
-	var data *storagepb.BidiWriteObjectRequest_ChecksummedData
-	if buf != nil {
-		data = &storagepb.BidiWriteObjectRequest_ChecksummedData{
-			ChecksummedData: &storagepb.ChecksummedData{
-				Content: buf,
-			},
-		}
-	}
-	req := &storagepb.BidiWriteObjectRequest{
-		Data:        data,
-		WriteOffset: offset,
-		FinishWrite: finishWrite,
-		Flush:       flush,
-		StateLookup: flush,
-	}
-	return req
-}
-
-type gRPCBidiWriteBufferSender interface {
-	// sendBuffer implementations should upload buf, respecting flush and
-	// finishWrite. Callers must guarantee that buf is not too long to fit in a
-	// gRPC message.
-	//
-	// If flush is true, implementations must not return until the data in buf is
-	// stable. If finishWrite is true, implementations must return the object on
-	// success.
-	sendBuffer(ctx context.Context, buf []byte, offset int64, flush, finishWrite bool) (*storagepb.Object, error)
-}
-
-type gRPCOneshotBidiWriteBufferSender struct {
-	firstMessage *storagepb.BidiWriteObjectRequest
-	raw          *gapic.Client
-	stream       storagepb.Storage_BidiWriteObjectClient
-	settings     *settings
-}
-
-func (w *gRPCWriter) newGRPCOneshotBidiWriteBufferSender() (*gRPCOneshotBidiWriteBufferSender, error) {
-	firstMessage := &storagepb.BidiWriteObjectRequest{
-		FirstMessage: &storagepb.BidiWriteObjectRequest_WriteObjectSpec{
-			WriteObjectSpec: w.spec,
-		},
-		CommonObjectRequestParams: toProtoCommonObjectRequestParams(w.encryptionKey),
-		// For a non-resumable upload, checksums must be sent in this message.
-		// TODO: Currently the checksums are only sent on the first message
-		// of the stream, but in the future, we must also support sending it
-		// on the *last* message of the stream (instead of the first).
-		ObjectChecksums: toProtoChecksums(w.sendCRC32C, w.attrs),
-	}
-
-	return &gRPCOneshotBidiWriteBufferSender{
-		firstMessage: firstMessage,
-		raw:          w.c.raw,
-		settings:     w.settings,
-	}, nil
-}
-
-func (s *gRPCOneshotBidiWriteBufferSender) sendBuffer(ctx context.Context, buf []byte, offset int64, flush, finishWrite bool) (obj *storagepb.Object, err error) {
-	var firstMessage *storagepb.BidiWriteObjectRequest
-	if s.stream == nil {
-		s.stream, err = s.raw.BidiWriteObject(ctx, s.settings.gax...)
-		if err != nil {
-			return
-		}
-		firstMessage = s.firstMessage
-	}
-	req := bidiWriteObjectRequest(buf, offset, flush, finishWrite)
-	if firstMessage != nil {
-		proto.Merge(req, firstMessage)
-	}
-
-	sendErr := s.stream.Send(req)
-	if sendErr != nil {
-		obj, err = drainInboundStream(s.stream)
-		s.stream = nil
-		if sendErr != io.EOF {
-			err = sendErr
-		}
-		return
-	}
-	// Oneshot uploads assume all flushes succeed
-
-	if finishWrite {
-		s.stream.CloseSend()
-		// Oneshot uploads only read from the response stream on completion or
-		// failure
-		obj, err = drainInboundStream(s.stream)
-		s.stream = nil
-		if err == io.EOF {
-			err = nil
-		}
-	}
-	return
-}
-
-type gRPCResumableBidiWriteBufferSender struct {
-	queryRetry        *retryConfig
-	upid              string
-	progress          func(int64)
-	raw               *gapic.Client
-	forceFirstMessage bool
-	stream            storagepb.Storage_BidiWriteObjectClient
-	flushOffset       int64
-	settings          *settings
-}
-
-func (w *gRPCWriter) newGRPCResumableBidiWriteBufferSender(ctx context.Context) (*gRPCResumableBidiWriteBufferSender, error) {
-	req := &storagepb.StartResumableWriteRequest{
-		WriteObjectSpec:           w.spec,
-		CommonObjectRequestParams: toProtoCommonObjectRequestParams(w.encryptionKey),
-		// TODO: Currently the checksums are only sent on the request to initialize
-		// the upload, but in the future, we must also support sending it
-		// on the *last* message of the stream.
-		ObjectChecksums: toProtoChecksums(w.sendCRC32C, w.attrs),
-	}
-
-	var upid string
-	err := run(ctx, func(ctx context.Context) error {
-		upres, err := w.c.raw.StartResumableWrite(ctx, req, w.settings.gax...)
-		upid = upres.GetUploadId()
-		return err
-	}, w.settings.retry, w.settings.idempotent)
-	if err != nil {
-		return nil, err
-	}
-
-	// Set up an initial connection for the 0 offset, so we don't query state
-	// unnecessarily for the first buffer. If we fail, we'll just retry in the
-	// normal connect path.
-	stream, err := w.c.raw.BidiWriteObject(ctx, w.settings.gax...)
-	if err != nil {
-		stream = nil
-	}
-
-	return &gRPCResumableBidiWriteBufferSender{
-		queryRetry:        w.settings.retry,
-		upid:              upid,
-		progress:          w.progress,
-		raw:               w.c.raw,
-		forceFirstMessage: true,
-		stream:            stream,
-		settings:          w.settings,
-	}, nil
-}
-
-// queryProgress is a helper that queries the status of the resumable upload
-// associated with the given upload ID.
-func (s *gRPCResumableBidiWriteBufferSender) queryProgress(ctx context.Context) (int64, error) {
-	var persistedSize int64
-	err := run(ctx, func(ctx context.Context) error {
-		q, err := s.raw.QueryWriteStatus(ctx, &storagepb.QueryWriteStatusRequest{
-			UploadId: s.upid,
-		}, s.settings.gax...)
-		// q.GetPersistedSize() will return 0 if q is nil.
-		persistedSize = q.GetPersistedSize()
-		return err
-	}, s.queryRetry, true)
-
-	return persistedSize, err
-}
-
-func (s *gRPCResumableBidiWriteBufferSender) sendBuffer(ctx context.Context, buf []byte, offset int64, flush, finishWrite bool) (obj *storagepb.Object, err error) {
-	if s.stream == nil {
-		// Determine offset and reconnect
-		s.flushOffset, err = s.queryProgress(ctx)
-		if err != nil {
-			return
-		}
-		s.stream, err = s.raw.BidiWriteObject(ctx, s.settings.gax...)
-		if err != nil {
-			return
-		}
-		s.forceFirstMessage = true
-	}
-
-	// clean up buf. We'll still write the message if a flush/finishWrite was
-	// requested.
-	if offset < s.flushOffset {
-		trim := s.flushOffset - offset
-		if int64(len(buf)) <= trim {
-			trim = int64(len(buf))
-		}
-		buf = buf[trim:]
-	}
-	if len(buf) == 0 && !flush && !finishWrite {
-		// no need to send anything
-		return nil, nil
-	}
-
-	req := bidiWriteObjectRequest(buf, offset, flush, finishWrite)
-	if s.forceFirstMessage {
-		req.FirstMessage = &storagepb.BidiWriteObjectRequest_UploadId{UploadId: s.upid}
-		s.forceFirstMessage = false
-	}
-
-	sendErr := s.stream.Send(req)
-	if sendErr != nil {
-		obj, err = drainInboundStream(s.stream)
-		s.stream = nil
-		if err == io.EOF {
-			// This is unexpected - we got an error on Send(), but not on Recv().
-			// Bubble up the sendErr.
-			err = sendErr
-		}
-		return
-	}
-
-	if finishWrite {
-		s.stream.CloseSend()
-		obj, err = drainInboundStream(s.stream)
-		s.stream = nil
-		if err == io.EOF {
-			err = nil
-			if obj.GetSize() > s.flushOffset {
-				s.progress(obj.GetSize())
-			}
-		}
-		return
-	}
-
-	if flush {
-		resp, err := s.stream.Recv()
-		if err != nil {
-			return nil, err
-		}
-		persistedOffset := resp.GetPersistedSize()
-		if persistedOffset > s.flushOffset {
-			s.flushOffset = persistedOffset
-			s.progress(s.flushOffset)
-		}
-	}
-	return
-}
-
-// uploadBuffer uploads the buffer at the given offset using a bi-directional
-// Write stream. It will open a new stream if necessary (on the first call or
-// after resuming from failure) and chunk the buffer per maxPerMessageWriteSize.
-// The final Object is returned on success if doneReading is true.
-//
-// Returns object and any error that is not retriable.
-func (w *gRPCWriter) uploadBuffer(ctx context.Context, recvd int, start int64, doneReading bool) (obj *storagepb.Object, err error) {
-	if w.streamSender == nil {
-		if w.append {
-			// Appendable object semantics
-			w.streamSender, err = w.newGRPCAppendableObjectBufferSender()
-		} else if doneReading || w.forceOneShot {
-			// One shot semantics
-			w.streamSender, err = w.newGRPCOneshotBidiWriteBufferSender()
-		} else {
-			// Resumable write semantics
-			w.streamSender, err = w.newGRPCResumableBidiWriteBufferSender(ctx)
-		}
-		if err != nil {
-			return
-		}
-	}
-
-	data := w.buf[:recvd]
-	offset := start
-	// We want to go through this loop at least once, in case we have to
-	// finishWrite with an empty buffer.
-	for {
-		// Send as much as we can fit into a single gRPC message. Only flush once,
-		// when sending the very last message.
-		l := maxPerMessageWriteSize
-		flush := false
-		if len(data) <= l {
-			l = len(data)
-			flush = true
-		}
-		obj, err = w.streamSender.sendBuffer(ctx, data[:l], offset, flush, flush && doneReading)
-		if err != nil {
-			return nil, err
-		}
-		data = data[l:]
-		offset += int64(l)
-		if len(data) == 0 {
-			// Update object size to match persisted offset.
-			if obj != nil {
-				obj.Size = offset
-			}
-			break
-		}
-	}
-	if w.flushInProgress {
-		w.setSize(offset)
-		w.flushInProgress = false
-		w.flushComplete <- offset
-	}
-	return
-}
-
-// read copies the data in the reader to the given buffer and reports how much
-// data was read into the buffer and if there is no more data to read (EOF).
-// read returns when either 1. the buffer is full, 2. Writer.Flush was called,
-// or 3. Writer.Close was called.
-func (w *gRPCWriter) read() (int, bool, error) {
-	// Set n to -1 to start the Read loop.
-	var n, recvd int = -1, 0
-	var err error
-	for err == nil && n != 0 {
-		// The routine blocks here until data is received.
-		n, err = w.reader.Read(w.buf[recvd:])
-		recvd += n
-	}
-	var done bool
-	if err == io.EOF {
-		err = nil
-		// EOF can come from Writer.Flush or Writer.Close.
-		if w.flushInProgress {
-			// Reset pipe for additional writes after the flush.
-			pr, pw := io.Pipe()
-			w.reader = pr
-			w.pw = pw
-			w.setPipeWriter(pw)
-		} else {
-			done = true
-		}
-	}
-	return recvd, done, err
-}
-
-// flush flushes the current buffer regardless of whether it is full or not.
-// It's the implementation for Writer.Flush.
-func (w *gRPCWriter) flush() (int64, error) {
-	if !w.append {
-		return 0, errors.New("Flush is supported only if Writer.Append is set to true")
-	}
-
-	// Close PipeWriter to trigger EOF on read side of the stream.
-	w.flushInProgress = true
-	w.pw.Close()
-
-	// Wait for flush to complete
-	offset := <-w.flushComplete
-	return offset, nil
-}
-
-func checkCanceled(err error) error {
-	if status.Code(err) == codes.Canceled {
-		return context.Canceled
-	}
-
-	return err
 }
